requirements.txt
transformers
torch
fastapi
uvicorn
datasets
evaluate
wandb
python-dotenv

